+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## dev
 
+* Fix a violated precondition in `result_buffer_to_string()` in newer versions
+  of `rustc`:
+    ```
+    unsafe precondition(s) violated: slice::get_unchecked requires that the
+    index is within the slice
+    ```
 * Explicitly specify the used Rust edition to silence compilation warnings.
 
 ## 0.6.0 (2023-11-14)